Refactor the way in which authentication is handled by the client and the server. Now, the server advertises what authentication methods it supports, and the client attempts to use each method in turn until one succeeds. Authentication methods are implemented by plug-ins on both sides of the communication.
This revision adds support for credentials forwarding on Windows using
SSPI. Currently, only SSPI on top of NTLM has been tested--with mixed
results. Windows users no longer have to enter a password to be
authenticated, but if the nodes running maestrod need to access
another computer to get access to other resources, the credentials
cannot be forwarded again because NTLM does not support this. There
may yet be a way for us to work around this, but at the moment, using
Kerberos for Windows authentication is likely to be the most
reliable way of handling this issue.

This change represents quite a lot of work done by Aron and me
together over the last few days. It is still not entirely complete,
but no functionality has been lost or broken. Even so, there is
still some polish to add (as always) aside from addressing the NTLM
credentials forwarding limitation. Documentation updates will follow
once the dust has settled.