Begin the Maestro /Programmer's Guide/, a document intended for people who will write code to extend Maestro. I started this document by pulling some text out of the /User's Guide/ that referred to low-level details of Maestro service and view plug-ins.
My intention with this document is to explain the following aspects of the
Maestro software as a whole:

   * How to write plug-ins (views, services, authorization clients, and
     authorization servers).
   * How the communication between client and server works.
   * The details of all signals sent between the client and the server.

Ultimately, someone should be able to write alternative implementations of
the Maestro client GUI that can take advantage of a subset of the features
and services offered by maestrod.

Right now, only the Launch Service is documented.