Added the ability to save individual stanzas. This implementation is a pretty messy hack because the current design of the Stanza Store, the Stanza Editor, and the Maestro GUI framework don't really support this capability well. Some major redesign work in the GUI is going to be needed at some point to address this and other issues.
This revision also includes initial support for the "Save Stanza As" action,
but it is currently disabled. The Stanza Editor does not handle that case
well at all, and it will take more work (hacks?) to get it behave correctly
when the possiblity exists for stanza files to be renamed on the fly.

This fixes Ticket #30, and it addresses #31 partially.